backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
PyPDF2==3.0.1
python-multipart==0.0.19
numpy==1.26.4
